sample_files/backend_template/requirements.txt
# Core Framework
fastapi==0.115.14
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.9  # For PostgreSQL
asyncpg==0.29.0  # Async PostgreSQL driver

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Environment & Configuration
python-dotenv==1.0.1
pydantic==2.11.7
pydantic-settings==2.7.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

# LLM & AI Framework
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
openai>=1.60.0

# Cache & Message Queue
redis==5.2.1

# Observability & Monitoring
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-instrumentation-sqlalchemy==0.49b2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-instrumentation-redis==0.49b2

# Graph Database
neo4j==5.20.0

# Task Queue
celery==5.4.0